Create HX711_v2.ino
Currently, there is an issue within the loop function for determining when to auto-tare. After initial auto-tare, the counter does not reset to zero, this can be fixed by removing the else keyword. However, when it is removed, it gets stuck auto-taring after every 3 measurements and cannot read the weight of an item on the scale.